# src/main/resources/configs/swagger.yaml
# as a demonstration of an API spec in YAML
swagger: '2.0'
info:
  title: WebKorhaz API
  description: WebKorhaz API
  version: "1.0.0"
# the domain of the service
host: localhost:8080
# array of all schemes that your API supports
schemes:
  - https
  - http
# will be prefixed to all paths
basePath: /api
produces:
  - application/json
paths:
  /login:
    post: 
      summary: Login method
      description: |
        Through the Login endpoint the user can login in the application. This endpoint if the login was succesful return a security token and the user's id
      parameters:
        - name: email
          in: body
          description: Login object
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      tags:
        - Login
      responses:
        200:
          description: Logged user's id and a security token
          schema:
              $ref: '#/definitions/LoginResponse'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /getDepartmentsAndTypes:
    get:
      summary: Get Departments and department's consultationHour's Types
      description: |
        The DepartmentsAndTypes endpoint returns information about the hospital's departments and about the consultation Hour's tpyes
      parameters: []
      tags:
        - Department
      responses:
        200:
          description: An array of products
          schema:
            type: array
            items:
              $ref: '#/definitions/Department'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /consultationHourSearch:
    post:
      summary: Search for ConsultationHour
      description: |
        The consultationHourSearch endpoint returns information about the searched consultationHours
      parameters:
      - name: request
        in: body
        description: Search data
        required: true
        schema:
          $ref: '#/definitions/ConsultationHourSearch'
      tags:
        - ConsultationHour
      responses:
        200:
          description: An array of products
          schema:
            type: array
            items:
              $ref: '#/definitions/ConsultationHour'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /appointment/list:
    get:
      summary: Get appointmentList
      description: |
        This endpoint returns information aboutthe user's appointment
      parameters:
      - name: userId
        in: query
        description: User's/Patient's id
        required: true
        type: integer
        format: int32
      tags:
        - Appointment
      responses:
        200:
          description: An array of products
          schema:
            type: array
            items:
              $ref: '#/definitions/Appointment'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /appointment:
    post:
      summary: Post appointment
      description: |
        With this endpoint user can create a new appointment
      parameters:
      - name: request
        in: body
        description: Appointment
        required: true
        schema:
          $ref: '#/definitions/Appointment'
      tags:
        - Appointment
      responses:
        200:
          description: Succes
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Put appointment
      description: |
        With this endpoint user can update an existing appointment
      parameters:
      - name: request
        in: body
        description: Appointment
        required: true
        schema:
          $ref: '#/definitions/Appointment'
      tags:
        - Appointment
      responses:
        200:
          description: Succes
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get appointment
      description: |
        With this endpoint user can get an existing appointment with it's id
      parameters:
      - name: appointmentId
        in: query
        description: Appointment
        required: true
        type: integer
        format: int32
      tags:
        - Appointment
      responses:
        200:
          description: Succes
          schema:
            $ref: '#/definitions/Appointment'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete appointment
      description: |
        With this endpoint user can delete an existing appointment
      parameters:
      - name: request
        in: body
        description: Appointment
        required: true
        schema:
          $ref: '#/definitions/Appointment'
      tags:
        - Appointment
      responses:
        200:
          description: Succes
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  LoginRequest:
    type: object
    properties:
      email:
        type: string
        description: Email addres
      passwordHash:
        type: string
        description: Email addres
  LoginResponse:
    type: object
    properties:
      userId:
        type: integer
        format: int32
        description: Id of the logged User
      token: 
        type : string
        description: user token
  Department:
    type: object
    properties:
      departmentId:
        type: integer
        format: int32
        description: Id of department
      departmentName:
        type: string
        description: Name of the department
      consultationHourType:
        type: array
        items:
          $ref: "#/definitions/ConsultationHourType"
  ConsultationHourType:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: Id of ConsultationHourType
      name:
        type: string
        description: Name of the ConsultationHourType
  ConsultationHourSearch:
    type: object
    properties:
      departmentName:
        type: string
        description: Department's name
      type:
        type: integer
        format: int32
        description: Type of the consultationHour
      beginDate:
        type: string
        format: date-time
        description: Start date of the ConsultationHour
      endDate:
        type: string
        format: date-time
        description: End date of the ConsultationHour
  ConsultationHour:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: Id of ConsultationHour
      typeId:
        type: integer
        format: int32
        description: Id of ConsultationHourType
      beginDate:
        type: string
        format: date-time
        description: Start date of the ConsultationHour
      endDate:
        type: string
        format: date-time
        description: End date of the ConsultationHour
      maxPatientCount:
        type: integer
        format: int32
        description: Maxinum number of patient count allowed in this consultationHour
      currentPatientCount:
        type: integer
        format: int32
        description: Current number of patient count in this consultationHour
  Appointment:
    type: object
    properties:
      appointmentId:
        type: integer
        format: int32
        description: Id of Appointment
      beginDate:
        type: string
        format: date-time
        description: Start date of the ConsultationHour
      endDate:
        type: string
        format: date-time
        description: End date of the ConsultationHour
      room:
        type: string
        description: Room of the ConsultationHour
      doctorsName:
        type: string
        description: Doctors name who goint to holf the ConsultationHour
      complaints:
        type: string
        description: Complaint of the patient
      userId:
        type: integer
        format: int32
        description: Id of the patient
      consultationHourId:
        type: integer
        format: int32
        description: Id of ConsultationHour
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string